backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-telegram-bot[job-queue]>=21.0
anthropic>=0.40.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
httpx>=0.27.0
